Abstracted endpoint authentication & database layer to core service
