+++ cart/api/cart.go
@@ -15,11 +15,11 @@ import (
 // ------------------------------------------------------
 // Setup Database
 
-// ProductDB instance.
-var ProductsDB = sqldb.Named("products")
+// Database instance for Plamatio Backend.
+var PlamatioDB = sqldb.Named("plamatio_db")
 
 // CartItemsTable instance.
-var CartItemsTable = &db.CartItemsTable{DB: ProductsDB}
+var CartItemsTable = &db.CartItemsTable{DB: PlamatioDB}
 
 // ------------------------------------------------------
 // Setup Caching
